convert_prisma

Access a PRISMA L1 HDF5 file and convert it to ENVI or GeoTiff format

convert_prisma(
  in_file,
  out_file,
  out_format = "ENVI",
  VNIR = TRUE,
  SWIR = TRUE,
  FULL = FALSE,
  source = "HCO",
  join_priority = "SWIR",
  ATCOR = TRUE,
  ATCOR_wls = "Nominal",
  PAN = TRUE,
  CLOUD = FALSE,
  LC = FALSE,
  GLINT = FALSE,
  overwrite = FALSE
)

Arguments

in_file
`character` full path of input HDF5 file

out_file
`character` full path of output file

out_format
`character`` ["TIF" | "ENVI"], Output format, Default: 'tif'

VNIR
`logical` if TRUE, create the VNIR image, Default: TRUE

SWIR
`logical` if TRUE, create the SWIR image, Default: TRUE

FULL
`logical` if TRUE, create a single multispectral image from VNIR and SWIR, Default: FALSE

source
`character` ["HC0" | "HRC"], Considered Data Cube Default: 'HRC'

join_priority
`character` ["VNIR" | "SWIR"], spectrometer to consider in the when join_spectra = TRUE, Default: SWIR - ignored if join_spectra is FALSE. Default: "VNIR"

ATCOR
logical` if TRUE, create the text files required to run ATCOR, Default: FALSE;

ATCOR_wls
`character`, or `numeric` If equal to `Nominal`, the only ATCOR wvl file created is the one containing Nominal wavelengths. If a numeric vector is provided, then one different wvl file is created for each selected COLUMN selected (e.g., if providing `ATCOR_wls = c(200, 800)`, then the wavelengths and FWHMs related to columns 200 and 800 are saved.)

PAN
`logical` if TRUE, also save the PAN data, default: TRUE

CLOUD
`logical` if TRUE, also save the cloud mask data, default: TRUE

LC
`logical` if TRUE, also save the land cover data, default: TRUE

GLINT
`logical` if TRUE, also save the GLINT mask data, default: TRUE

overwrite
`logical` if TRUE, existing files are overwritten, default: FALSE

Value

The function is called for its side effects

Examples

if (FALSE) {
 in_file  <- "/home/lb/tmp/test/PRS_L1_STD_OFFL_20190825103112_20190825103117_0001.he5"
 out_file <- "/home/lb/tmp/test/test_1"
 out_format <- "ENVI"

 # Save a full image, prioritizing the VNIR spectrometer and save in ENVI format
 convert_prisma(in_file       = in_file,
                out_file      = out_file,
                out_format    = out_format,
                FULL          = TRUE,
                join_priority = "VNIR",
                LC            = TRUE,
                CLOUD         = TRUE,
                overwrite     = TRUE
                )

 # Save a full image, prioritizing the SWIR spectrometer and save in ENVI format,
 # Also create ATCOR files, with Nominal wavelengths, and those for columns
 # 200 and 800 of the cube
 convert_prisma(in_file       = in_file,
                out_file      = out_file,
                out_format    = out_format,
                FULL          = TRUE,
                join_priority = "VNIR",
                ATCOR         = TRUE,
                ATCOR_wls     = c(200,800),
                LC            = TRUE,
                CLOUD         = TRUE,
                overwrite     = TRUE
                )

}