prisma_create_additional

helper function used to process and save additional data sets such as CLOUD, LC and GLINT

prisma_create_additional(
  f,
  type,
  out_file,
  out_format,
  base_georef,
  fill_gaps,
  fix_geo,
  in_L2_file = NULL
)

Arguments

f
input data he5 from caller

type
`character` type of dataset to be created ("CLD", "LC" or "GlNT")

out_file
output file name for the dataset

out_format
`character`` ["TIF" | "ENVI"], Output format, Default: 'tif'

base_georef
`logical` if TRUE, apply base georeferencing on L1, L2B/C data, using the "Georeference from input GLT" procedure explained here: https://www.harrisgeospatial.com/docs/backgroundgltbowtiecorrection.html, Default: TRUE

fill_gaps
`logical` if TRUE, when georeferencing on L1, L2B/C data, substitute missing values with results of a 3x3 focal filter on the georeferenced data, Default: TRUE

fix_geo
`logical` if TRUE, apply a 90 metres correction to georeferencing info of l2D data, Default: FALSE

in_L2_file
`character` full path of an L2 file to be used to extract better georeeferencing info and angles for a corresponding L1 file. If not NULL, and `in_file` is a L1 file, the LAT and LON fields used for bowtie georeferencing are taken from the L2 file instead than from the L1 file

Value

The function is called for its side effects