prisma_make_atcor

helper function used to create accessory files needed for ATCOR atmospheric correction

prisma_make_atcor(
  f,
  out_file,
  ATCOR_wls,
  wls,
  fwhms,
  order_vnir,
  order_swir,
  join_priority,
  source
)

Arguments

f
input data he5 from caller

out_file
main prismaread output file name - passed by caller

ATCOR_wls
NULL, or `numeric` If NULL the only ATCOR wvl file created is the one containing Nominal wavelengths. If a numeric vector is provided, then one different wvl file is created for each selected COLUMN selected (e.g., if providing `ATCOR_wls = c(200, 800)`, then the wavelengths and FWHMs related to columns 200 and 800 are saved.)

wls
passed by caller - array of PRISMA wavelengths

fwhms
passed by caller - array of PRISMA fwhms

order_vnir
passed by caller - ordering of array of PRISMA VNIR wavelengths

order_swir
passed by caller - ordering of array of PRISMA SWIR wavelengths

join_priority
`character` ["VNIR" | "SWIR"], spectrometer to consider in the when join_spectra = TRUE, Default: SWIR - ignored if join_spectra is FALSE. Default: "VNIR"

source
`character` ["HC0" | "HRC"], Considered Data Cube Default: 'HCO'

Value

`list` containing required info according to `proc_lev`